Extracted API endpoints methods as service classes
